feat(angular): prevent overwriting existing tailwind config
